+ Protect from exceptions when rendering tooltip functions. + unrelated clean-up to slim down Column file
