+++ cmp/grid/impl/Utils.ts
@@ -4,12 +4,11 @@
  *
  * Copyright © 2023 Extremely Heavy Industries Inc.
  */
-import {ColumnRenderer, GroupRowRenderer} from '@xh/hoist/cmp/grid';
-import {isFunction} from 'lodash';
+import {Column, ColumnGroup, ColumnRenderer, GroupRowRenderer} from '@xh/hoist/cmp/grid';
+import {HeaderClassParams} from '@xh/hoist/kit/ag-grid';
+import {castArray, isFunction} from 'lodash';
 
-/**
- * @internal
- */
+/** @internal */
 export function managedRenderer<T extends ColumnRenderer | GroupRowRenderer>(
     fn: T,
     identifier: string
@@ -23,4 +22,39 @@ export function managedRenderer<T extends ColumnRenderer | GroupRowRenderer>(
             return '#ERROR';
         }
     } as unknown as T;
+}
+
+/**
+ *   Generate CSS classes for headers.
+ *   Default alignment classes are mixed in with any provided custom classes.
+ *
+ *   @internal
+ */
+export function getAgHeaderClassFn(
+    column: Column | ColumnGroup
+): (params: HeaderClassParams) => string[] {
+    const {headerClass, headerAlign, gridModel} = column;
+
+    return agParams => {
+        let r = [];
+        if (headerClass) {
+            r = castArray(
+                isFunction(headerClass) ? headerClass({column, gridModel, agParams}) : headerClass
+            );
+        }
+
+        if (headerAlign === 'center' || headerAlign === 'right') {
+            r.push('xh-column-header-align-' + headerAlign);
+        }
+
+        if (column instanceof Column && column.isTreeColumn && column.headerHasExpandCollapse) {
+            r.push('xh-column-header--with-expand-collapse');
+        }
+
+        if (gridModel.headerMenuDisplay === 'hover') {
+            r.push('xh-column-header--hoverable');
+        }
+
+        return r;
+    };
 }
